Cree y termine la vista, el css y el controlador de Editar Usuario y termine esa funcionalidad

+++ src/main/java/RetailManagementSystem/dominio/puertos/RepositorioUsuario.java
@@ -15,7 +15,7 @@ public interface RepositorioUsuario {
 
     Optional<Usuario> obtenerUsuarioPorEmail(String email);
 
-    Usuario obtenerUsuarioPorId(int idUsuario);
+    Usuario obtenerUsuarioPorId(Long idUsuario);
 
     List<Usuario> obtenerTodosLosUsuarios();
